+++ Changelog.md
@@ -11,6 +11,7 @@ based on merged pull requests. Search GitHub issues and pull requests for smalle
 - Add null checks to avoid NPE [#612](https://github.com/entur/lamassu/pull/612)
 - Remove redundant entity cache interfaces [#613](https://github.com/entur/lamassu/pull/613)
 - Add request-scope cache for reading entities [#619](https://github.com/entur/lamassu/pull/619)
+- Only filter out vehicles that are assigned to non-virtual stations [#626](https://github.com/entur/lamassu/pull/626)
   [](AUTOMATIC_CHANGELOG_PLACEHOLDER_DO_NOT_REMOVE)
 
 ## 1.1.0